tweak file unique name pattern
in practice, i still want the full date in file names. this helps when an import
or a search result spans multiple days.

the md5 portion doesn't need to be full length. 24 chars is more than enough to
deter fuskers, and is also more than enough to prevent collisions in file data
(especially paired with the full datetime).